Rename system to systems in api and add post mock
